Day2/Exercise/requirements.txt
Django==4.2.0
psycopg2-binary==2.9.6
Pillow==10.0.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0

